NBDapi/requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.26.0
autopep8==2.0.0
click==8.1.3
colorama==0.4.5
dictdiffer==0.9.0
fastapi==0.85.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
loguru==0.6.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
pycodestyle==2.10.0
pydantic==1.10.2
pypika-tortoise==0.1.6
python-multipart==0.0.5
pytz==2022.5
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tomli==2.0.1
tomlkit==0.11.5
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.19.0
